+++ neo/framework/FileSystem.cpp
@@ -1828,7 +1828,7 @@ int idFileSystemLocal::GetFileList( const char *relativePath, const idStrList &e
 		return 0;
 	}
 
-	int pathLength = strlen( relativePath );
+	int pathLength = (int)strlen( relativePath );
 	if ( pathLength ) {
 		pathLength++;	// for the trailing '/'
 	}
@@ -2239,7 +2239,7 @@ void idFileSystemLocal::TouchFileList_f( const idCmdArgs &args ) {
 	const char *buffer = NULL;
 	idParser src( LEXFL_NOFATALERRORS | LEXFL_NOSTRINGCONCAT | LEXFL_ALLOWMULTICHARLITERALS | LEXFL_ALLOWBACKSLASHSTRINGCONCAT );
 	if ( fileSystem->ReadFile( args.Argv( 1 ), ( void** )&buffer, NULL ) && buffer ) {
-		src.LoadMemory( buffer, strlen( buffer ), args.Argv( 1 ) );
+		src.LoadMemory( buffer, (int)strlen( buffer ), args.Argv( 1 ) );
 		if ( src.IsLoaded() ) {
 			idToken token;
 			while( src.ReadToken( &token ) ) {
